+++ AWSAuthSDK/Sources/AWSMobileClient/AWSMobileClient.swift
@@ -121,7 +121,11 @@ final public class AWSMobileClient: _AWSMobileClient {
     
     /// The registered listeners who want to observe change in `UserState`.
     var listeners: [(AnyObject, UserStateChangeCallback)] = []
-    
+
+    /// Holds a reference to the AWSInfo configuration object
+    internal lazy var awsInfo: AWSInfo = {
+        return AWSInfo.default()
+    }()
     
     // MARK: Public API variables
     
@@ -141,16 +145,11 @@ final public class AWSMobileClient: _AWSMobileClient {
         return self.internalCredentialsProvider?.identityId
     }
     
-    @objc override public func interceptApplication(_ application: UIApplication, open url: URL, sourceApplication: String?, annotation: Any) -> Bool {
-        
-        return _AWSMobileClient.sharedInstance().interceptApplication(application, open: url, sourceApplication: sourceApplication, annotation: annotation)
-    }
-    
     /// Returns true if there is a user currently signed in.
     @objc public var isSignedIn: Bool {
         get {
             if (operateInLegacyMode) {
-                return _AWSMobileClient.sharedInstance().isLoggedIn
+                return self.isLoggedIn
             } else {
                 return self.cachedLoginsMap.count > 0
             }
@@ -160,11 +159,30 @@ final public class AWSMobileClient: _AWSMobileClient {
     
     /// The singleton instance of `AWSMobileClient`.
     ///
+    /// **Deprecation note:** use `AWSMobileClient.default()` since it communicates better the API intent.
+    ///
     /// - Returns: The singleton `AWSMobileClient` instance.
+    @available(*, deprecated, renamed: "default")
     @objc override public class func sharedInstance() -> AWSMobileClient {
+        return self.default()
+    }
+
+    /// The default instance of `AWSMobileClient`. The configuration is loaded from the `awsconfiguration.json` file.
+    ///
+    /// You can manage your own instance of `AWSMobileClient` by constructing it with `AWSMobileClient(configuration)`,
+    /// however please note that multiple instances of AWSMobileClient is not supported.
+    ///
+    /// **Implementation Notes:**
+    ///
+    /// The `AWSMobileClient` relies on `AWSInfo` for configuration and once that class is initialized, the configuration
+    /// cannot be reset and/or re-initialized. Therefore, even though you can instantiate `AWSMobileClient` multiple times,
+    /// all instances will have the same configuration reference.
+    ///
+    /// - Returns: The default `AWSMobileClient` instance
+    @objc public class func `default`() -> AWSMobileClient {
         return _sharedInstance
     }
-    
+
     public func handleAuthResponse(_ application: UIApplication, open url: URL, sourceApplication: String?, annotation: Any) {
         if (isCognitoAuthRegistered) {
             AWSCognitoAuth.init(forKey: CognitoAuthRegistrationKey).application(application, open: url, options: [:])
@@ -207,7 +225,7 @@ final public class AWSMobileClient: _AWSMobileClient {
             if self.federationProvider == .hostedUI {
                 loadOAuthURIQueryParametersFromKeychain()
                 
-                let infoDictionaryMobileClient = AWSInfo.default().rootInfoDictionary["Auth"] as? [String: [String: Any]]
+                let infoDictionaryMobileClient = self.awsInfo.rootInfoDictionary["Auth"] as? [String: [String: Any]]
                 var infoDictionary: [String: Any]? = infoDictionaryMobileClient?["Default"]?["OAuth"] as? [String: Any]
                 
                 let clientId = infoDictionary?["AppClientId"] as? String
@@ -266,9 +284,9 @@ final public class AWSMobileClient: _AWSMobileClient {
                 
                 self.isAuthorizationAvailable = true
                 self.internalCredentialsProvider = credentialsProvider
-                _AWSMobileClient.sharedInstance().update(self)
+                self.update(self)
                 self.internalCredentialsProvider?.setIdentityProviderManagerOnce(self)
-                _AWSMobileClient.sharedInstance().registerConfigSignInProviders()
+                self.registerConfigSignInProviders()
                 
                if (self.internalCredentialsProvider?.identityId != nil) {
                     if (federationProvider == .none) {
@@ -471,7 +489,7 @@ final public class AWSMobileClient: _AWSMobileClient {
             }
             
         } else {
-            _AWSMobileClient.sharedInstance().showSign(inScreen: navigationController, signInUIConfiguration: signInUIOptions, completionHandler: { providerName, token, error in
+            self.showSign(inScreen: navigationController, signInUIConfiguration: signInUIOptions, completionHandler: { providerName, token, error in
                 if error == nil {
                     if (providerName == IdentityProvider.facebook.rawValue) || (providerName == IdentityProvider.google.rawValue) {
                         self.federatedSignIn(providerName: providerName!, token: token!, completionHandler: completionHandler)