+++ AWSAuthSDK/Tests/AWSMobileClientTests/AWSMobileClientTests.swift
@@ -22,7 +22,7 @@ class AWSMobileClientTests: AWSMobileClientBaseTests {
         signUpUser(username: username)
 
         let verificationCodeSent = expectation(description: "verification code should be sent via email.")
-        AWSMobileClient.sharedInstance().resendSignUpCode(username: username) { (result, error) in
+        AWSMobileClient.default().resendSignUpCode(username: username) { (result, error) in
             if let error = error {
                 XCTFail("Failed due to error: \(error.localizedDescription)")
                 return
@@ -52,7 +52,7 @@ class AWSMobileClientTests: AWSMobileClientBaseTests {
         signUpAndVerifyUser(username: username)
         signIn(username: username)
         let tokensExpectation = expectation(description: "Successfully fetch AWS Credentials")
-        AWSMobileClient.sharedInstance().getTokens { (tokens, error) in
+        AWSMobileClient.default().getTokens { (tokens, error) in
             if let tokens = tokens {
                 XCTAssertNotNil(tokens.idToken)
             } else if let error = error {
@@ -86,12 +86,12 @@ class AWSMobileClientTests: AWSMobileClientBaseTests {
             return
         }
         
-        AWSMobileClient.sharedInstance().federatedSignIn(providerName: IdentityProvider.developer.rawValue, token: token!, federatedSignInOptions: FederatedSignInOptions(cognitoIdentityId: identityId!)) { (userState, error) in
+        AWSMobileClient.default().federatedSignIn(providerName: IdentityProvider.developer.rawValue, token: token!, federatedSignInOptions: FederatedSignInOptions(cognitoIdentityId: identityId!)) { (userState, error) in
             XCTAssertNil(error, "Expected successful federation.")
         }
         
         let credentialsExpectation = expectation(description: "Successfully fetch AWS Credentials")
-        AWSMobileClient.sharedInstance().getAWSCredentials { (credentials, error) in
+        AWSMobileClient.default().getAWSCredentials { (credentials, error) in
             if let credentials = credentials {
                 XCTAssertNotNil(credentials.accessKey)
                 XCTAssertNotNil(credentials.secretKey)
@@ -102,17 +102,17 @@ class AWSMobileClientTests: AWSMobileClientBaseTests {
         }
         wait(for: [credentialsExpectation], timeout: 10)
         
-        AWSMobileClient.sharedInstance().signOut()
+        AWSMobileClient.default().signOut()
         
-        XCTAssertFalse(AWSMobileClient.sharedInstance().isSignedIn, "User should be signed out.")
-        XCTAssertNil(AWSMobileClient.sharedInstance().identityId, "Identity Id should be nil after signing out.")
+        XCTAssertFalse(AWSMobileClient.default().isSignedIn, "User should be signed out.")
+        XCTAssertNil(AWSMobileClient.default().identityId, "Identity Id should be nil after signing out.")
         
         let username = "testUser" + UUID().uuidString
         signUpAndVerifyUser(username: username)
         signIn(username: username)
         
         let credentialsExpectation2 = expectation(description: "Successfully fetch AWS Credentials")
-        AWSMobileClient.sharedInstance().getAWSCredentials { (credentials, error) in
+        AWSMobileClient.default().getAWSCredentials { (credentials, error) in
             if let credentials = credentials {
                 XCTAssertNotNil(credentials.accessKey)
                 XCTAssertNotNil(credentials.secretKey)
@@ -123,14 +123,14 @@ class AWSMobileClientTests: AWSMobileClientBaseTests {
         }
         wait(for: [credentialsExpectation2], timeout: 10)
         
-        AWSMobileClient.sharedInstance().signOut()
+        AWSMobileClient.default().signOut()
     }
     
     func testUserStateNotifications() {
         var signInExpectation = false
         var signOutExpectation = false
-        XCTAssertTrue(AWSMobileClient.sharedInstance().listeners.count == 1, "Expecting only 1 listener.")
-        AWSMobileClient.sharedInstance().addUserStateListener(self) { (userState, info) in
+        XCTAssertTrue(AWSMobileClient.default().listeners.count == 1, "Expecting only 1 listener.")
+        AWSMobileClient.default().addUserStateListener(self) { (userState, info) in
             switch(userState) {
             case .signedIn:
                 signInExpectation = true
@@ -144,20 +144,20 @@ class AWSMobileClientTests: AWSMobileClientBaseTests {
         signUpAndVerifyUser(username: username)
         signIn(username: username)
         sleep(1)
-        AWSMobileClient.sharedInstance().signOut()
+        AWSMobileClient.default().signOut()
         sleep(1)
         XCTAssertTrue(signInExpectation, "Expected Sign In to be true")
         XCTAssertTrue(signOutExpectation, "Expected Sign Out to be true")
-        XCTAssertTrue(AWSMobileClient.sharedInstance().listeners.count == 2, "Expecting 2 listeners.")
-        AWSMobileClient.sharedInstance().removeUserStateListener(self)
-        XCTAssertTrue(AWSMobileClient.sharedInstance().listeners.count == 1, "Expecting only 1 listener.")
+        XCTAssertTrue(AWSMobileClient.default().listeners.count == 2, "Expecting 2 listeners.")
+        AWSMobileClient.default().removeUserStateListener(self)
+        XCTAssertTrue(AWSMobileClient.default().listeners.count == 1, "Expecting only 1 listener.")
     }
     
     func testGetIdentityId() {
-        XCTAssertNil(AWSMobileClient.sharedInstance().identityId, "Identity Id should be nil after initialize.")
+        XCTAssertNil(AWSMobileClient.default().identityId, "Identity Id should be nil after initialize.")
 
         let identityIdExpectation = expectation(description: "Request to GetIdentityID is complete")
-        AWSMobileClient.sharedInstance().getIdentityId().continueWith(block: { (task) -> Any? in
+        AWSMobileClient.default().getIdentityId().continueWith(block: { (task) -> Any? in
             XCTAssertNil(task.error)
             XCTAssertNotNil(task.result, "GetIdentityId should not return nil.")
             identityIdExpectation.fulfill()
@@ -165,7 +165,7 @@ class AWSMobileClientTests: AWSMobileClientBaseTests {
         })
         wait(for: [identityIdExpectation], timeout: 5)
 
-        XCTAssertNotNil(AWSMobileClient.sharedInstance().identityId, "Identity Id should not be nil.")
+        XCTAssertNotNil(AWSMobileClient.default().identityId, "Identity Id should not be nil.")
     }
     
     func testMultipleGetIdentityId() {